requirements.txt
annotated-types==0.7.0
anyio==4.8.0
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.6
h11==0.14.0
idna==3.10
mypy==1.14.1
mypy-extensions==1.0.0
pydantic==2.10.5
pydantic_core==2.27.2
sniffio==1.3.1
starlette==0.41.3
tomli==2.2.1
typing_extensions==4.12.2
uvicorn==0.34.0
pytest==7.1.2        # Pytest 버전은 상황에 맞춰 조정
httpx==0.24.1        # Starlette TestClient에 필요한 라이브러리
sqlalchemy==2.0.20   # ORM 사용에 필요한 라이브러리 (버전은 예시)
python-dotenv==0.19.1